resolve model referencing error for review
