+++ user/models.py
@@ -3,6 +3,8 @@ from django.contrib.auth.models import AbstractUser
 from restaurant.models import Restaurant, Categories
 from phonenumber_field.modelfields import PhoneNumberField
 
+from datetime import datetime
+
 
 class DineSafelyUser(AbstractUser):
     favorite_restaurants = models.ManyToManyField(Restaurant, blank=True)
@@ -90,3 +92,33 @@ class User_Profile(models.Model):
 
     def __str__(self):
         return f"{self.user.username}_user_profile"
+
+
+class Review(models.Model):
+    # Basic info
+    user = models.ForeignKey(
+        DineSafelyUser, on_delete=models.CASCADE, related_name="reviews"
+    )
+    restaurant = models.ForeignKey(
+        Restaurant, on_delete=models.CASCADE, related_name="reviews"
+    )
+    time = models.DateTimeField(default=datetime.now, editable=False)
+    content = models.TextField()
+
+    # Ratings
+    rating = models.PositiveIntegerField(default=0)
+    rating_safety = models.PositiveIntegerField(default=0)
+    # Accessibility Ratings
+    rating_entry = models.PositiveIntegerField(default=0)
+    rating_door = models.PositiveIntegerField(default=0)
+    rating_table = models.PositiveIntegerField(default=0)
+    rating_bathroom = models.PositiveIntegerField(default=0)
+    rating_path = models.PositiveIntegerField(default=0)
+
+    # Images
+    image1 = models.ImageField(null=True, blank=True, upload_to="review_images/")
+    image2 = models.ImageField(null=True, blank=True, upload_to="review_images/")
+    image3 = models.ImageField(null=True, blank=True, upload_to="review_images/")
+
+    def __str__(self):
+        return f"{self.user.username} review on {self.restaurant.restaurant_name}"
